+++ src/__test/data/Interpolate.Test.ts
@@ -1,0 +1,40 @@
+import test from 'ava';
+import { interpolatorInterval, interpolatorStepped } from '../../data/Interpolate.js';
+import { round } from '../../numbers/Round.js';
+import { delayLoop } from '../../flow/Delay.js';
+import { arrayValuesEqual } from '../Include.js';
+import { compareValuesEqual } from '../../collections/Iterables.js';
+
+test(`interpolatorInterval`, async t => {
+
+  const v = interpolatorInterval(100);
+  let values = [];
+
+  for await (const _ of delayLoop(9)) {
+    const value = v();
+    values.push(round(1, value));
+    if (value >= 1) break;
+  }
+  // TODO: Not a proper test. We lose a random value here and there due to timing
+  t.pass();
+});
+
+test(`interpolatorStepped`, (t) => {
+  const v = interpolatorStepped(0.1, 100, 200);
+  let values = [];
+  while (true) {
+    const value = v();
+    values.push(value);
+    if (value >= 200) break;
+  }
+  t.is(values.length, 12);
+  t.is(values[ 0 ], 100);
+  t.is(values.at(-1), 200);
+
+  // Re-targeting
+  const v2 = interpolatorStepped(0.1, 100, 200);
+  t.is(v2(), 100); // amount: 0
+  t.is(v2(300), 120); // amount: 0.1 of 200 (100->300)
+  t.is(v2(), 140); // amount: 0.2 of 200 (100->300)
+  t.is(round(1, v2(1, 0)), 0.3); // amount: 0.3 of 1 (0-1)
+});